+++ neurom/features/__init__.py
@@ -37,40 +37,84 @@ Examples:
     >>> ax_sec_len = features.get('section_lengths', nrn, neurite_type=neurom.AXON)
 """
 
+from enum import Enum
 import numpy as np
 
+from neurom.core import Population, Neuron, Neurite
+from neurom.core.neuron import iter_neurites
+from neurom.core.types import NeuriteType, tree_type_checker as is_type
 from neurom.exceptions import NeuroMError
 
-NEURITEFEATURES = dict()
-NEURONFEATURES = dict()
+_NEURITE_FEATURES = dict()
+_NEURON_FEATURES = dict()
+_POPULATION_FEATURES = dict()
 
 
-def _find_feature_func(feature_name):
-    """Returns the python function used when getting a feature with `neurom.get(feature_name)`."""
-    feat = NEURITEFEATURES.get(feature_name) or NEURONFEATURES.get(feature_name)
-    if feat is None:
-        raise NeuroMError(f'Unable to find feature: {feature_name}')
-    return feat
+class NameSpace(Enum):
+    """The level of morphology abstraction that feature applies to."""
+    NEURITE = 'neurite'
+    NEURON = 'neuron'
+    POPULATION = 'population'
 
 
-def _get_feature_value_and_func(feature_name, obj, **kwargs):
+def _get_feature_value_and_func(feature_name, obj, agg=None, **kwargs):
     """Obtain a feature from a set of morphology objects.
 
     Arguments:
-        feature(string): feature to extract
-        obj: a neuron, population or neurite tree
+        feature_name(string): feature to extract
+        obj (Neurite|Neuron|Population): neurite, neuron or population
+        agg (string|None):
         kwargs: parameters to forward to underlying worker functions
 
     Returns:
         A tuple (feature, func) of the feature value and its function
     """
-    feat = _find_feature_func(feature_name)
+    is_obj_list = isinstance(obj, (list, tuple))
+    if not isinstance(obj, (Neurite, Neuron, Population)) and not is_obj_list:
+        raise NeuroMError('Only Neurite, Neuron, Population or list, tuple of Neurite, Neuron can'
+                          ' be used for feature calculation')
+    if agg is not None and not hasattr(np, agg):
+        raise NeuroMError('`agg` argument must an aggregating function of numpy package.')
 
-    res = feat(obj, **kwargs)
-    if len(feat.shape) != 0:
-        res = np.array(list(res))
+    neurite_filter = is_type(kwargs.get('neurite_type', NeuriteType.all))
+    res, feature_ = None, None
 
-    return res, feat
+    if isinstance(obj, Neurite) or (is_obj_list and isinstance(obj[0], Neurite)):
+        # input is a neurite or a list of neurites
+        if feature_name in _NEURITE_FEATURES:
+            feature_ = _NEURITE_FEATURES[feature_name]
+            if isinstance(obj, Neurite):
+                res = feature_(obj, **kwargs)
+            else:
+                res = [feature_(s, **kwargs) for s in obj]
+    elif isinstance(obj, Neuron):
+        # input is a neuron
+        if feature_name in _NEURON_FEATURES:
+            feature_ = _NEURON_FEATURES[feature_name]
+            res = feature_(obj, **kwargs)
+        elif feature_name in _NEURITE_FEATURES:
+            feature_ = _NEURITE_FEATURES[feature_name]
+            res = sum(feature_(s, **kwargs) for s in iter_neurites(obj, filt=neurite_filter))
+    elif isinstance(obj, Population) or (is_obj_list and isinstance(obj[0], Neuron)):
+        # input is a neuron population or a list of neurons
+        if feature_name in _POPULATION_FEATURES:
+            feature_ = _POPULATION_FEATURES[feature_name]
+            res = feature_(obj, **kwargs)
+        elif feature_name in _NEURON_FEATURES:
+            feature_ = _NEURON_FEATURES[feature_name]
+            res = [feature_(n, **kwargs) for n in obj]
+        elif feature_name in _NEURITE_FEATURES:
+            feature_ = _NEURITE_FEATURES[feature_name]
+            res = [sum(feature_(s, **kwargs) for s in iter_neurites(n, filt=neurite_filter))
+                   for n in obj]
+
+    if res is None or feature_ is None:
+        raise NeuroMError(f'Cant apply "{feature_name}" feature. Please check that it exists, '
+                          'and can be applied to your input. See the features documentation page.')
+    if isinstance(res, list) and agg is not None:
+        res = getattr(np, agg)(res, axis=0)
+
+    return res, feature_
 
 
 def get(feature_name, obj, **kwargs):
@@ -91,34 +135,32 @@ def get(feature_name, obj, **kwargs):
     return _get_feature_value_and_func(feature_name, obj, **kwargs)[0]
 
 
-def _register_feature(namespace, name, func, shape):
+def _register_feature(namespace: NameSpace, name, func, shape):
     """Register a feature to be applied.
 
     Upon registration, an attribute 'shape' containing the expected
     shape of the function return is added to 'func'.
 
     Arguments:
-        namespace(string): a namespace (must be 'NEURITEFEATURES' or 'NEURONFEATURES')
+        namespace(string): a namespace, see :class:`NameSpace`
         name(string): name of the feature, used to access the feature via `neurom.features.get()`.
         func(callable): single parameter function of a neurite.
         shape(tuple): the expected shape of the feature values
     """
     setattr(func, 'shape', shape)
+    setattr(func, 'namespace', namespace)
 
-    assert namespace in {'NEURITEFEATURES', 'NEURONFEATURES'}
-    feature_dict = globals()[namespace]
-
-    if name in feature_dict:
+    if name in _FEATURES:
         raise NeuroMError('Attempt to hide registered feature %s' % name)
-    feature_dict[name] = func
+    _FEATURES[name] = func
 
 
-def feature(shape, namespace=None, name=None):
+def feature(shape, namespace: NameSpace, name=None):
     """Feature decorator to automatically register the feature in the appropriate namespace.
 
     Arguments:
         shape(tuple): the expected shape of the feature values
-        namespace(string): a namespace (must be 'NEURITEFEATURES' or 'NEURONFEATURES')
+        namespace(string): a namespace, see :class:`NameSpace`
         name(string): name of the feature, used to access the feature via `neurom.features.get()`.
     """
     def inner(func):